refactor(host-application): split build orchestrator into focused submodules
