+++ apps/desktop/src-tauri/crates/host-application/src/app_service/build_orchestrator/build_run_control.rs
@@ -1,0 +1,91 @@
+use super::super::{emit_event, terminate_child_process, AppService, RunEmitter};
+use super::BuildResponseAction;
+use anyhow::{anyhow, Result};
+use host_domain::{now_rfc3339, RunEvent, RunState, TaskStatus};
+
+impl AppService {
+    pub fn build_respond(
+        &self,
+        run_id: &str,
+        action: BuildResponseAction,
+        payload: Option<&str>,
+        emitter: RunEmitter,
+    ) -> Result<bool> {
+        let mut runs = self
+            .runs
+            .lock()
+            .map_err(|_| anyhow!("Run state lock poisoned"))?;
+        let run = runs
+            .get_mut(run_id)
+            .ok_or_else(|| anyhow!("Run not found: {run_id}"))?;
+
+        match action {
+            BuildResponseAction::Approve => {
+                if payload
+                    .map(|entry| entry.contains("git push"))
+                    .unwrap_or(false)
+                {
+                    run.summary.last_message =
+                        Some("Approved sensitive command: git push".to_string());
+                } else {
+                    run.summary.last_message = Some("Approval received".to_string());
+                }
+                run.summary.state = RunState::Running;
+            }
+            BuildResponseAction::Deny => {
+                run.summary.last_message = Some("Command denied by user".to_string());
+                run.summary.state = RunState::Blocked;
+                let _ = self.task_transition(
+                    &run.repo_path,
+                    &run.task_id,
+                    TaskStatus::Blocked,
+                    Some("User denied command"),
+                );
+            }
+            BuildResponseAction::Message => {
+                run.summary.last_message = payload.map(|entry| entry.to_string());
+            }
+        }
+
+        emit_event(
+            &emitter,
+            RunEvent::AgentThought {
+                run_id: run_id.to_string(),
+                message: run
+                    .summary
+                    .last_message
+                    .clone()
+                    .unwrap_or_else(|| "User response applied".to_string()),
+                timestamp: now_rfc3339(),
+            },
+        );
+
+        Ok(true)
+    }
+
+    pub fn build_stop(&self, run_id: &str, emitter: RunEmitter) -> Result<bool> {
+        let mut runs = self
+            .runs
+            .lock()
+            .map_err(|_| anyhow!("Run state lock poisoned"))?;
+        let run = runs
+            .get_mut(run_id)
+            .ok_or_else(|| anyhow!("Run not found: {run_id}"))?;
+
+        terminate_child_process(&mut run.child);
+        run.summary.state = RunState::Stopped;
+        run.summary.last_message = Some("Run stopped by user".to_string());
+
+        emit_event(
+            &emitter,
+            RunEvent::RunFinished {
+                run_id: run_id.to_string(),
+                message: "Run stopped".to_string(),
+                timestamp: now_rfc3339(),
+                success: false,
+            },
+        );
+
+        Ok(true)
+    }
+}
